!1185 [pytorch][dev][BertNER] DTS2022062509691已修复，去掉规避代码 !1185 [pytorch][dev][BertNER] DTS2022062509691已修复，去掉规避代码

+++ PyTorch/dev/nlp/BertNER_ID3438_for_PyTorch/models/layers/crf.py
@@ -219,7 +219,7 @@ class CRF(nn.Module):
         # shape: (batch_size,)
         seq_ends = mask.long().sum(dim=0) - 1
         # shape: (batch_size,) DTS2022062509691
-        last_tags = tags[seq_ends.long(), bs_index]
+        last_tags = tags[seq_ends, bs_index]
         # shape: (batch_size,)
         score += self.end_transitions[last_tags]
